BE/Fix/#178 : 토큰값 수정

+++ Backend/src/main/java/ssafy/d210/backend/dto/response/certificate/CertificateToken.java
@@ -9,5 +9,5 @@ import java.math.BigInteger;
 @Builder
 public class CertificateToken {
 
-    private BigInteger token;
+    private int token;
 }
